requirements.txt
Django
djangorestframework
django-cors-headers
drf-yasg
pylint
psycopg2
bcrypt==3.2.0